Changed gene_indices to long*, allows for larger number of genes

+++ gm_creature.h
@@ -29,7 +29,7 @@ typedef struct Gene {
 //a creature is a collection of genes
 //it contains an array of ints which are the indices of the genes it contains
 typedef struct Creature {
-    int* gene_indices;
+    long* gene_indices;
     int num_genes;
 } Creature;
